MatrixCalculations
In the following program I took two 3x3 matrices A and B (The values are taken just for example and can be changed as per requirement ) , then did addition and subtraction between them. I also used inheritance concept for my own convenience.

You can refer to the comments in the program to understand the code a little better.