MatrixCalculations
In the following program I took two 3x3 matrices A, B (The values are taken as an example and can be modified based on your requirements ) and did addition as well as subtraction operations between them. I also used inheritance concept just for my convenience.

You can refer to the comments I made to understand the program a little better.